Refine landing page for SRP and performance
