+++ apps/landing/src/components/MotionBackdrop.tsx
@@ -1,0 +1,21 @@
+export function MotionBackdrop() {
+  return (
+    <div className="motion-backdrop" aria-hidden="true">
+      <div className="motion-grid" />
+      <div className="motion-scan motion-scan-primary" />
+      <div className="motion-scan motion-scan-secondary" />
+      <div className="motion-data-field">
+        <span />
+        <span />
+        <span />
+        <span />
+        <span />
+        <span />
+        <span />
+        <span />
+        <span />
+        <span />
+      </div>
+    </div>
+  );
+}
