+++ apps/landing/src/components/WorkflowCarousel.tsx
@@ -1,0 +1,52 @@
+import { useState } from 'react';
+import { workflowSteps } from '../content/workflowSteps';
+import { WorkflowPreview } from './WorkflowPreview';
+
+type WorkflowStepId = (typeof workflowSteps)[number]['id'];
+
+export function WorkflowCarousel({ prefersReducedMotion }: { prefersReducedMotion: boolean }) {
+  const [activeWorkflowStep, setActiveWorkflowStep] = useState<WorkflowStepId>('prompt');
+  const activeWorkflow = workflowSteps.find((step) => step.id === activeWorkflowStep) ?? workflowSteps[0];
+
+  const advanceWorkflowStep = () => {
+    if (prefersReducedMotion) {
+      return;
+    }
+
+    setActiveWorkflowStep((currentStep) => {
+      const currentIndex = workflowSteps.findIndex((step) => step.id === currentStep);
+      const nextIndex = currentIndex === -1 ? 0 : (currentIndex + 1) % workflowSteps.length;
+      return workflowSteps[nextIndex]!.id;
+    });
+  };
+
+  return (
+    <div id="workflow" className="workflow-carousel" aria-label="Feature workflow carousel">
+      <div className="workflow-copy">
+        <p className="eyebrow">Seamless workflow</p>
+        <h2>{activeWorkflow.title}</h2>
+        <p>{activeWorkflow.copy}</p>
+      </div>
+      <div className="workflow-tabs" role="tablist" aria-label="Choose workflow demo">
+        {workflowSteps.map(({ id, icon: Icon, title }) => (
+          <button
+            key={id}
+            type="button"
+            role="tab"
+            aria-selected={activeWorkflowStep === id}
+            className={activeWorkflowStep === id ? 'active' : undefined}
+            onClick={() => setActiveWorkflowStep(id)}
+          >
+            <Icon size={16} aria-hidden="true" />
+            {title}
+          </button>
+        ))}
+      </div>
+      <WorkflowPreview
+        activeStep={activeWorkflowStep}
+        onDemoEnded={advanceWorkflowStep}
+        prefersReducedMotion={prefersReducedMotion}
+      />
+    </div>
+  );
+}
