+++ apps/landing/src/content/editorProof.ts
@@ -1,0 +1,6 @@
+export const editorProof = [
+  'Layered slide canvas',
+  'Pages, thumbnails, and drag reordering',
+  'Text toolbar with translate actions',
+  'PNG export and local folder persistence',
+] as const;
