+++ apps/landing/src/content/externalLinks.ts
@@ -1,0 +1,7 @@
+export const externalLinks = {
+  githubRepo: 'https://github.com/ErickWendel/localstudio',
+  chromeBuiltInAi:
+    'https://developer.chrome.com/docs/ai/built-in?utm_source=localstudio.dev&utm_medium=referral&utm_campaign=localstudio_thanks',
+  huggingFaceWebMl:
+    'https://huggingface.co/webml-community?utm_source=localstudio.dev&utm_medium=referral&utm_campaign=localstudio_thanks',
+} as const;
