+++ apps/landing/src/content/featureMediaImages.ts
@@ -1,0 +1,27 @@
+export const featureMediaImages = {
+  prompt: {
+    src: '/prompt-to-slide-showcase.png',
+    alt: 'LocalStudio prompt-to-slide editor with an AI Design Revolution slide',
+    aspectRatio: '2982 / 2390',
+  },
+  image: {
+    src: '/prompt-to-image-showcase.png',
+    alt: 'LocalStudio prompt-to-image editor with a generated AI chip image on a slide',
+    aspectRatio: '2970 / 2398',
+  },
+  translate: {
+    src: '/translate-showcase.png',
+    alt: 'LocalStudio translate editor showing the AI Design Revolution slide translated into Portuguese',
+    aspectRatio: '2980 / 2396',
+  },
+  edit: {
+    src: '/edit-images-showcase.png',
+    alt: 'LocalStudio edit images workflow showing background removal segmentation on an image layer',
+    aspectRatio: '2982 / 2396',
+  },
+  local: {
+    src: '/project-history-showcase.png',
+    alt: 'LocalStudio project history panel showing saved local versions',
+    aspectRatio: '4110 / 2402',
+  },
+} as const;
