+++ apps/landing/src/content/socialLinks.ts
@@ -1,0 +1,14 @@
+export const socialLinks = [
+  {
+    label: 'YouTube',
+    href: 'https://www.youtube.com/@ErickWendelAcademy',
+  },
+  {
+    label: 'GitHub',
+    href: 'https://github.com/erickWendel',
+  },
+  {
+    label: 'LinkedIn',
+    href: 'https://www.linkedin.com/in/erickwendel/',
+  },
+] as const;
