+++ apps/landing/src/styles/base.css
@@ -1,0 +1,160 @@
+* {
+  box-sizing: border-box;
+}
+
+html {
+  scroll-behavior: smooth;
+}
+
+body {
+  margin: 0;
+  min-width: 0;
+  min-height: 100vh;
+  background: var(--ls-bg);
+  color: var(--ls-text);
+  font-family: var(--ls-font-body);
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+button,
+a {
+  -webkit-tap-highlight-color: transparent;
+}
+
+.landing-shell {
+  position: relative;
+  overflow: hidden;
+}
+
+.hero-section,
+.demo-section,
+.showcase-section,
+.web-ai-section,
+.webmcp-section,
+.features-section,
+.requirements-section,
+.thanks-section,
+.closing-section {
+  position: relative;
+  z-index: 1;
+}
+
+.eyebrow {
+  margin: 0 0 14px;
+  color: var(--ls-green-fixed);
+  font-family: var(--ls-font-display);
+  font-size: 12px;
+  font-weight: 900;
+  letter-spacing: 0.16em;
+  text-transform: uppercase;
+}
+
+h1,
+h2,
+h3,
+p {
+  margin-top: 0;
+}
+
+h1,
+h2,
+h3 {
+  letter-spacing: 0;
+}
+
+h1 {
+  max-width: 620px;
+  margin-bottom: 18px;
+  font-family: var(--ls-font-display);
+  font-size: clamp(36px, 4.4vw, 60px);
+  font-weight: 900;
+  line-height: 1;
+}
+
+h2 {
+  margin-bottom: 16px;
+  font-family: var(--ls-font-display);
+  font-size: clamp(30px, 4vw, 56px);
+  font-weight: 900;
+  line-height: 1.02;
+}
+
+h3 {
+  margin-bottom: 10px;
+  font-family: var(--ls-font-display);
+  font-size: 18px;
+  font-weight: 900;
+}
+
+.hero-subtitle,
+.section-heading p,
+.web-ai-copy p,
+.feature-panel p {
+  color: var(--ls-muted);
+  font-size: clamp(16px, 1.4vw, 19px);
+  line-height: 1.58;
+}
+
+.header-cta,
+.primary-action,
+.secondary-action,
+.inline-section-link {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  border-radius: 6px;
+  font-weight: 900;
+  transition:
+    border-color 150ms ease,
+    background 150ms ease,
+    color 150ms ease,
+    transform 150ms ease,
+    box-shadow 150ms ease;
+}
+
+.primary-action,
+.secondary-action {
+  min-height: 52px;
+  padding: 0 22px;
+  border: 1px solid rgba(55, 253, 118, 0.62);
+}
+
+.primary-action {
+  background: var(--ls-green);
+  color: var(--ls-on-primary);
+}
+
+.secondary-action {
+  background: rgba(0, 0, 0, 0.48);
+  color: var(--ls-text);
+}
+
+.header-cta:hover,
+.primary-action:hover,
+.secondary-action:hover,
+.inline-section-link:hover,
+.inline-section-link:focus-visible {
+  transform: translateY(-2px);
+  box-shadow: 0 0 26px rgba(55, 253, 118, 0.24);
+}
+
+.showcase-section,
+.demo-section,
+.web-ai-section,
+.webmcp-section,
+.features-section,
+.requirements-section,
+.thanks-section,
+.closing-section {
+  padding: clamp(54px, 8vw, 104px) clamp(18px, 5vw, 72px);
+}
+
+.section-heading {
+  max-width: 780px;
+  margin-bottom: 34px;
+}
